server/requirements.txt
Flask
SQLAlchemy 
Flask-SQLAlchemy
Flask-Cors
flask_cors
mysqlclient
python-dotenv
flask-JWT-Extended
flask-uploads
cloudinary